fix: remove support for prefix, add new export for prefixed version
This change removes the default support for the `prefix` option in the Tailwind CSS configuration.
Instead, a new export is added to the `package.json` file, allowing users to import a prefixed
version of the CSS if needed. This fixes issues related to the `--spacing` function.

BREAKING CHANGE: The default export no longer supports the prefix option. Users who need prefixed
classes should import `tw-animate-css/prefix`.